+++ core-ng-mongo/src/main/java/core/framework/mongo/module/MongoConfig.java
@@ -29,7 +29,7 @@ public class MongoConfig extends Config {
         this.name = name;
 
         var mongo = new MongoImpl();
-        this.context.startupHook.add(mongo::initialize);
+        this.context.startupHook.initialize.add(mongo::initialize);
         this.context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> mongo.close());
         context.beanFactory.bind(Mongo.class, name, mongo);
         this.mongo = mongo;