+++ core-ng/src/main/java/core/framework/module/CacheConfig.java
@@ -130,7 +130,7 @@ public class CacheConfig extends Config {
             logger.info("create redis local cache store");
             LocalCacheStore localCache = localCacheStore();
             var thread = new RedisSubscribeThread("cache-invalidator", redis, new InvalidateLocalCacheMessageListener(localCache), RedisLocalCacheStore.CHANNEL_INVALIDATE_CACHE);
-            context.startupHook.add(thread::start);
+            context.startupHook.start.add(thread::start);
             context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> thread.close());
             redisLocalCacheStore = new RedisLocalCacheStore(localCache, redisCacheStore, redis);
         }